fix: change some order props and fix delivery method not getting its values on the get order by id method

+++ backend/WebStore.API/WebStore.Application/DTOs/BasketDto/BasketUpdateDto.cs
@@ -9,6 +9,6 @@ public record BasketUpdateDto(
     string Id,
     [Required] List<BasketItemDto> BasketItems,
     string DeliveryMethodId,
-    decimal ShippingPrice,
-    decimal TotalPrice
+    double ShippingPrice,
+    double TotalPrice
     );
